add feature tests
