+++ test/features/user_sees_event_list_links_test.rb
