+++ test/features/user_sees_events_by_hour_test.rb
